adding 2 test scripts in the separate directory
